+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+# 1.0.0-rc.5
+
 - CLI/Option change:
   - Add `apex-insert-final-newline` option ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/83)).
 - Remove newline before SOQL expression in ForInit ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/28)).